llamaIndex/pyproject.toml
[project]
name = "llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crawl4ai>=0.8.6",
    "dotenv-python>=0.0.1",
    "llama-index-core>=0.14.20",
    "llama-index-embeddings-huggingface>=0.7.0",
    "llama-index-llms-groq>=0.5.0",
    "llama-index-llms-huggingface>=0.7.0",
    "llama-index-readers-web>=0.6.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
